+++ lib/widgets/slides/slide_preview.dart
@@ -17,6 +17,7 @@ import 'mermaid_diagram.dart';
 import 'chart_hover.dart';
 import 'video_playhead_bus.dart';
 import 'previews/callout_overlay.dart';
+import 'previews/retrying_image.dart';
 import '../../l10n/app_localizations.dart';
 import '../../models/chart.dart';
 import '../../services/improvement/canvas_layout.dart';